requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
hiredis==2.2.3
httplib2==0.22.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
typer==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
